+++ server/src/main/java/com/project/chamong/place/service/VisitedPlaceService.java
@@ -1,7 +1,12 @@
 package com.project.chamong.place.service;
 
+import com.project.chamong.auth.dto.AuthorizedMemberDto;
+import com.project.chamong.camping.entity.Content;
+import com.project.chamong.camping.service.CampingApiService;
+import com.project.chamong.exception.BusinessLogicException;
+import com.project.chamong.exception.ExceptionCode;
 import com.project.chamong.member.entity.Member;
-import com.project.chamong.member.repository.MemberRepository;
+import com.project.chamong.member.service.MemberService;
 import com.project.chamong.place.dto.VisitedPlaceDto;
 import com.project.chamong.place.entity.VisitedPlace;
 import com.project.chamong.place.mapper.VisitedPlaceMapper;
@@ -11,43 +16,60 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class VisitedPlaceService {
     private final VisitedPlaceRepository visitedPlaceRepository;
-    private final VisitedPlaceMapper visitedPlaceMapper;
-    private final MemberRepository memberRepository;
+    private final MemberService memberService;
+    private final CampingApiService campingApiService;
+    private final VisitedPlaceMapper mapper;
 
-    public List<VisitedPlaceDto.Response> findAll(){
-        List<VisitedPlace> visitedPlaces = visitedPlaceRepository.findAll();
-        return visitedPlaces.stream()
-                .map(visitedPlaceMapper::visitedPlaceToResponse)
-                .collect(Collectors.toList());
+    public List< VisitedPlaceDto.Response> findVisitedPlaces(){
+        List<VisitedPlace> visitedPlaces = visitedPlaceRepository.findWithContentAll();
+        
+        return mapper.visitedPlacesToResponseDtos(visitedPlaces);
     }
-    public Optional<VisitedPlaceDto.Response> findById(Long id){
-        Optional<VisitedPlace> visitedPlace = visitedPlaceRepository.findById(id);
-        return visitedPlace.map(visitedPlaceMapper::visitedPlaceToResponse);
+    
+    public VisitedPlaceDto.Response saveVisitedPlace(Long contentId, AuthorizedMemberDto authorizedMemberDto){
+        Member findMember = memberService.findById(authorizedMemberDto.getId());
+    
+        Content findContent = verifyVisitedPlaceExist(findMember, contentId);
+    
+        VisitedPlaceDto.Post postDto = VisitedPlaceDto.Post.builder().content(findContent).build();
+    
+        VisitedPlace visitedPlace = VisitedPlace.createVisitedPlace(postDto);
+    
+        visitedPlace.setMember(findMember);
+    
+        visitedPlaceRepository.save(visitedPlace);
+        
+        return mapper.visitedPlaceToResponseDto(visitedPlace);
     }
-    public VisitedPlaceDto.Response create(VisitedPlaceDto.Post postDto){
-        Member member = memberRepository.findById(postDto.getMemberId())
-                .orElseThrow(()-> new IllegalArgumentException("Member not found with ID: "+ postDto.getMemberId()));
-        VisitedPlace visitedPlace = VisitedPlace.createVisitedPlace(postDto, member);
-        visitedPlace.setMember(member);
 
-        return visitedPlaceMapper.visitedPlaceToResponse(visitedPlaceRepository.save(visitedPlace));
-    }
-    public VisitedPlaceDto.Response update(Long id, VisitedPlaceDto.Patch patchDto){
-        VisitedPlace visitedPlace = visitedPlaceRepository.findById(id)
-                .orElseThrow(()-> new IllegalArgumentException("VisitedPlace not found with ID: "+ id));
-        visitedPlace.update(patchDto);
-        return visitedPlaceMapper.visitedPlaceToResponse(visitedPlace);
-    }
-    public void delete(Long id){
+    public void deleteVisitedPlace(Long id, AuthorizedMemberDto authorizedMemberDto){
+        VisitedPlace findVisitedPlace = verifyVisitedPlaceExist(id);
+        
+        if(findVisitedPlace.getMember().getId().equals(authorizedMemberDto.getId())){
+            throw new BusinessLogicException(ExceptionCode.VISITED_PLACE_DELETE_NO_PERMISSION);
+        }
+        
         visitedPlaceRepository.deleteById(id);
+    }
+    public Content verifyVisitedPlaceExist(Member member, Long contentId){
+        VisitedPlace findVisitedPlace = visitedPlaceRepository.findByMemberIdAndContentId(member.getId(), contentId);
+        
+        if(findVisitedPlace != null){
+            throw new BusinessLogicException(ExceptionCode.VISITED_PLACE_EXISTS);
+        }
+        
+        return campingApiService.findContent(contentId);
+    }
+    
+    public VisitedPlace verifyVisitedPlaceExist(Long id){
+        return visitedPlaceRepository.findById(id)
+          .orElseThrow(() -> new BusinessLogicException(ExceptionCode.VISITED_PLACE_NOT_FOUND));
     }
 
 }
